+++ tortoise/contrib/test/__init__.py
@@ -1,20 +1,22 @@
-import asyncio as _asyncio
+import asyncio
 import os as _os
-from copy import deepcopy
-from typing import List
+from asyncio.selector_events import BaseSelectorEventLoop
+from typing import List, Optional
 from unittest import SkipTest, expectedFailure, skip, skipIf, skipUnless  # noqa
 
 from asynctest import TestCase as _TestCase
 from asynctest import _fail_on
+from asynctest.case import _Policy
 
-from tortoise import Tortoise
+from tortoise import ContextVar, Tortoise
 from tortoise.backends.base.config_generator import generate_config as _generate_config
 from tortoise.exceptions import DBConnectionError
-from tortoise.transactions import start_transaction
+from tortoise.transactions import current_transaction_map, start_transaction
 
 __all__ = ('SimpleTestCase', 'IsolatedTestCase', 'TestCase', 'SkipTest', 'expectedFailure',
-           'skip', 'skipIf', 'skipUnless', 'initializer', 'finalizer')
-_TORTOISE_TEST_DB = _os.environ.get('TORTOISE_TEST_DB', 'sqlite:///tmp/test-{}.sqlite')
+           'skip', 'skipIf', 'skipUnless', 'env_initializer', 'initializer', 'finalizer',
+           'getDBConfig')
+_TORTOISE_TEST_DB = 'sqlite://:memory:'
 
 expectedFailure.__doc__ = """
 Mark test as expecting failiure.
@@ -23,13 +25,18 @@ On success it will be marked as unexpected success.
 """
 
 _CONFIG = {}  # type: dict
-_APPS = {}  # type: dict
 _CONNECTIONS = {}  # type: dict
+_SELECTOR = None  # type: ignore
+_LOOP = None  # type: BaseSelectorEventLoop
+_MODULES = []  # type: List[str]
 
 
 def getDBConfig(app_label: str, modules: List[str]) -> dict:
     """
     DB Config factory, for use in testing.
+
+    :param app_label: Label of the app (must be distinct for multiple apps).
+    :param modules: List of modules to look for models in.
     """
     return _generate_config(
         _TORTOISE_TEST_DB,
@@ -52,38 +59,81 @@ async def _init_db(config):
     await Tortoise.generate_schemas()
 
 
-def initializer():
+def _restore_default():
+    Tortoise.apps = {}
+    Tortoise._connections = _CONNECTIONS.copy()
+    for name in Tortoise._connections.keys():
+        current_transaction_map[name] = ContextVar(name, default=None)
+    Tortoise._init_apps(_CONFIG['apps'])
+    Tortoise._inited = True
+
+
+def initializer(modules: List[str], db_url: Optional[str] = None,
+                loop: Optional[BaseSelectorEventLoop] = None) -> None:
     """
     Sets up the DB for testing. Must be called as part of test environment setup.
+
+    :param modules: List of modules to look for models in.
+    :param db_url: The db_url, defaults to ``sqlite://:memory``.
+    :param loop: Optional event loop.
     """
     # pylint: disable=W0603
     global _CONFIG
-    global _APPS
     global _CONNECTIONS
+    global _SELECTOR
+    global _LOOP
+    global _TORTOISE_TEST_DB
+    global _MODULES
+    _MODULES = modules
+    if db_url is not None:
+        _TORTOISE_TEST_DB = db_url
     _CONFIG = getDBConfig(
         app_label='models',
-        modules=['tortoise.tests.testmodels'],
+        modules=_MODULES,
     )
 
-    loop = _asyncio.get_event_loop()
+    loop = loop or asyncio.get_event_loop()
+    _LOOP = loop
+    _SELECTOR = loop._selector  # type: ignore
     loop.run_until_complete(_init_db(_CONFIG))
-    _APPS = deepcopy(Tortoise.apps)
     _CONNECTIONS = Tortoise._connections.copy()
-    loop.run_until_complete(Tortoise._reset_connections())
+    Tortoise.apps = {}
+    Tortoise._connections = {}
+    Tortoise._inited = False
 
 
-def finalizer():
+def finalizer() -> None:
     """
     Cleans up the DB after testing. Must be called as part of the test environment teardown.
     """
-    Tortoise.apps = deepcopy(_APPS)
-    Tortoise._connections = _CONNECTIONS.copy()
-    loop = _asyncio.get_event_loop()
+    _restore_default()
+    loop = _LOOP
+    loop._selector = _SELECTOR
     loop.run_until_complete(Tortoise._drop_databases())
+
+
+def env_initializer() -> None:
+    """
+    Calls ``initializer()`` with parameters mapped from environment variables.
+
+    ``TORTOISE_TEST_MODULES``:
+        A comma-separated list of modules to include *(required)*
+    ``TORTOISE_TEST_DB``:
+        The db_url of the test db. *(optional*)
+    """
+    modules = str(_os.environ.get('TORTOISE_TEST_MODULES', 'tortoise.tests.testmodels')).split(',')
+    db_url = _os.environ.get('TORTOISE_TEST_DB', 'sqlite://:memory:')
+    if not modules:
+        raise Exception('TORTOISE_TEST_MODULES envvar not defined')
+    initializer(modules, db_url=db_url)
 
 
 class SimpleTestCase(_TestCase):
     """
+    The Tortoise base test class.
+
+    This will ensure that your DB environment has a test double set up for use.
+
     An asyncio capable test class that provides some helper functions.
 
     Will run any ``test_*()`` function either as sync or async, depending
@@ -92,6 +142,21 @@ class SimpleTestCase(_TestCase):
 
     Based on `asynctest <http://asynctest.readthedocs.io/>`_
     """
+    use_default_loop = True
+
+    def _init_loop(self):
+        if self.use_default_loop:
+            self.loop = _LOOP
+            loop = None
+        else:
+            loop = self.loop = asyncio.new_event_loop()
+
+        policy = _Policy(asyncio.get_event_loop_policy(),
+                         loop, self.forbid_get_event_loop)
+
+        asyncio.set_event_loop_policy(policy)
+
+        self.loop = self._patch_loop(self.loop)
 
     async def _setUpDB(self):
         pass
@@ -109,7 +174,7 @@ class SimpleTestCase(_TestCase):
         self._checker.before_test(self)
 
         self.loop.run_until_complete(self._setUpDB())
-        if _asyncio.iscoroutinefunction(self.setUp):
+        if asyncio.iscoroutinefunction(self.setUp):
             self.loop.run_until_complete(self.setUp())
         else:
             self.setUp()
@@ -118,11 +183,15 @@ class SimpleTestCase(_TestCase):
         self.loop._asynctest_ran = False
 
     def _tearDown(self) -> None:
-        self.loop.run_until_complete(self._tearDownDB())
-        if _asyncio.iscoroutinefunction(self.tearDown):
+        if asyncio.iscoroutinefunction(self.tearDown):
             self.loop.run_until_complete(self.tearDown())
         else:
             self.tearDown()
+        self.loop.run_until_complete(self._tearDownDB())
+        Tortoise.apps = {}
+        Tortoise._connections = {}
+        Tortoise._inited = False
+        current_transaction_map.clear()
 
         # post-test checks
         self._checker.check_test(self)
@@ -144,12 +213,16 @@ class IsolatedTestCase(SimpleTestCase):
     async def _setUpDB(self):
         config = getDBConfig(
             app_label='models',
-            modules=['tortoise.tests.testmodels'],
+            modules=_MODULES,
         )
         await Tortoise.init(config, _create_db=True)
         await Tortoise.generate_schemas()
+        self._connections = Tortoise._connections.copy()
 
     async def _tearDownDB(self) -> None:
+        Tortoise._connections = self._connections.copy()
+        for name in Tortoise._connections.keys():
+            current_transaction_map[name] = ContextVar(name, default=None)
         await Tortoise._drop_databases()
 
 
@@ -160,13 +233,9 @@ class TestCase(SimpleTestCase):
     """
 
     async def _setUpDB(self):
-        Tortoise.apps = deepcopy(_APPS)
-        Tortoise._connections = _CONNECTIONS.copy()
-        await Tortoise.init(_CONFIG)
-
+        _restore_default()
         self.transaction = await start_transaction()  # pylint: disable=W0201
 
     async def _tearDownDB(self) -> None:
+        _restore_default()
         await self.transaction.rollback()
-        # Have to reset connections because tests are run in different loops
-        await Tortoise._reset_connections()